// cms/src/store/ducks/messages/types.ts
export interface IdName {
  id: number;
  name: string;
}

interface User {
  id: number;
  name: string;
  email: string;
  attachment: Attachment[];
}

export interface Attachment {
  name: string;
  id: number;
  attachmentType: AttachmentsType;
}

export interface BaseMessages {
  id: number;
  dateTimeLastResponse: Date;
  subject: string;
  unread: number;
  type: IdName;
  status: IdName;
  from: User;
  recipient: User;
  idParentMessage: number;
  messageAttachments: Attachment[];
  reply: Reply[];
}

interface AttachmentsType {
  id: number;
  name: string;
  mime: {
    media: string;
    suffix: string;
  };
}

interface Reply {
  id: number;
  dateTime: Date;
  text: string;
  replyFrom: User;
}
// ...
export type Messages = BaseMessages;